+++ db/blob/blob_log_writer.h
@@ -4,8 +4,6 @@
 //  (found in the LICENSE.Apache file in the root directory).
 #pragma once
 
-#ifndef ROCKSDB_LITE
-
 #include <cstdint>
 #include <memory>
 #include <string>
@@ -82,4 +80,3 @@ class BlobLogWriter {
 };
 
 }  // namespace ROCKSDB_NAMESPACE
-#endif  // ROCKSDB_LITE